PySlideView: now using 64 bits integers

+++ bridge/cxx/include/bhxx/Runtime.hpp
@@ -102,9 +102,9 @@ public:
      * @step_delay   The delay of the changes (used for nested loops)
      */
     template <typename T>
-    void slide_view(BhArray<T>* view_ptr, size_t dim, int slide, int shape_change,
-                    int view_shape, int view_stride, int step_delay) {
-        bh_slide_dim d{(int64_t)dim, slide, shape_change, view_stride, view_shape, step_delay};
+    void slide_view(BhArray<T>* view_ptr, int64_t dim, int64_t slide, int64_t shape_change,
+                    int64_t view_shape, int64_t view_stride, int64_t step_delay) {
+        bh_slide_dim d{dim, slide, shape_change, view_stride, view_shape, step_delay};
         view_ptr->slides.dims.push_back(d);
     }
 
@@ -117,7 +117,7 @@ public:
      * @reset_it The amount of iterations before a reset
      */
     template <typename T>
-    void add_reset(BhArray<T>* view_ptr, size_t dim, size_t reset_it) {
+    void add_reset(BhArray<T>* view_ptr, int64_t dim, int64_t reset_it) {
         view_ptr->slides.resets.insert({dim, reset_it});
         view_ptr->slides.changes_since_reset.insert({dim, 0});
     }
